+++ CHANGELOG.md
@@ -14,6 +14,7 @@ These notes are publicly posted in [production](https://wordplay.dev/updates), s
 
 - We fixed the text cursor position when spaces aren't shown explicitly.
 - We made the full range of emojis visible in the editor's emoji chooser.
+- We now only show conflicts related to the current source in the conflict list.
 
 ## 0.17.2 - 2026-03-21